% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\alias{plot_matrix.list}
\alias{plot_matrix.data.frame}
\alias{print.plotmatrix}
\title{A scatterplot matrix with densities on the diagonal.}
\usage{
plot_matrix(data, ...)

\method{plot_matrix}{list}(
  data,
  marg.exp = 0.02,
  labels = names(data),
  alpha = 1,
  pch = 16,
  size = unit(0.03, "npc"),
  ...
)

\method{plot_matrix}{data.frame}(
  data,
  marg.exp = 0.02,
  labels = names(data),
  alpha = 1,
  pch = 16,
  size = unit(0.03, "npc"),
  ...
)

\method{print}{plotmatrix}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{data}{Data to plot.}

\item{\dots}{optional arguments, passed to \code{\link{hist}}.}

\item{marg.exp}{Fraction by which to expand the plot at the margins.}

\item{labels}{Names of variables plotted.}

\item{alpha, pch, size}{Refer to the plotted points in the scatterplots.}

\item{x}{\code{plot_matrix} object to display.}

\item{newpage}{logical; if \code{TRUE}, \code{grid.newpage()} will be called
before the graphics are drawn.}

\item{vp}{viewport to use.  See \code{\link[grid]{viewport}}.}
}
\description{
A special scatterplot matrix.
}
\examples{
\donttest{ # requires dplyr
  library(dplyr)

  data.frame(
    a=rexp(n=1000,rate=1/3),
    b=rnorm(1000)
  ) |>
    mutate(
      c=a+b^2,
      d=a-b^3
    ) -> x

  print(plot_matrix(x,alpha=0.2))

  g <- plot_matrix(
    x[-2],
    labels=c(
      expression(alpha),
      expression(beta),
      expression(phi)
    ),
    alpha=0.3
  )
  print(g)

  print(plot_matrix(as.list(x),alpha=0.2,breaks="scott"))
}
}
