% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{circumstance-package}
\alias{circumstance-package}
\alias{circumstance,package}
\title{circumstance package}
\description{
\pkg{circumstance} provides tools for parallelizing certain \pkg{pomp} calculations.
}
