% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue.R, R/mif2.R, R/pmcmc.R
\name{continue}
\alias{continue}
\alias{continue,missing-method}
\alias{continue,ANY-method}
\alias{continue,mif2List-method}
\alias{continue,pmcmcList-method}
\title{Continue an iterative calculation}
\usage{
continue(object, ...)

\S4method{continue}{mif2List}(object, ..., .options.future = list(seed = TRUE))

\S4method{continue}{pmcmcList}(object, ..., .options.future = list(seed = TRUE))
}
\arguments{
\item{object}{the result of an iterative \pkg{pomp} computation}

\item{...}{additional arguments will be passed to the underlying method.
This allows one to modify parameters used in the original computations.}

\item{.options.future}{list of options for the \pkg{doFuture} backend.
See \code{\link[doFuture]{\%dofuture\%}} for details.}
}
\description{
Continue an iterative computation where it left off.
}
\seealso{
\code{\link{mif2}}
}
