% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mif2.R
\name{mif2}
\alias{mif2}
\alias{mif2,ANY,data.frame-method}
\alias{mif2,ANY,missing-method}
\alias{mif2,pompList,missing-method}
\alias{mif2,pfilterList,missing-method}
\alias{mif2,mif2List,missing-method}
\title{Parallel iterated filtering}
\usage{
\S4method{mif2}{ANY,data.frame}(data, starts, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)

\S4method{mif2}{ANY,missing}(data, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)

\S4method{mif2}{pompList,missing}(data, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)

\S4method{mif2}{pfilterList,missing}(data, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)

\S4method{mif2}{mif2List,missing}(data, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)
}
\arguments{
\item{data}{passed to \code{\link[pomp:mif2]{pomp::mif2}}}

\item{starts}{data frame containing parameters at which to begin iterated filtering}

\item{...}{all additional arguments are passed to \code{\link[pomp:mif2]{pomp::mif2}}}

\item{seed}{seed for the parallel random-number generator.
Setting \code{seed=TRUE} passes the current seed to the parallel RNG.
See \code{\link[doFuture]{\%dofuture\%}} for more information.}

\item{chunk.size}{average number of elements processed per chunk.
See \code{\link[doFuture]{\%dofuture\%}} for more information.}

\item{scheduling}{average number of chunks that each worker processes.
See \code{\link[doFuture]{\%dofuture\%}} for more information.
This is ignored unless \code{chunk.size=NULL}.}
}
\description{
Runs multiple instances of \code{mif2} using \code{foreach}.
}
\seealso{
\code{\link[pomp:mif2]{pomp::mif2}}.
}
