% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\docType{methods}
\name{pfilter}
\alias{pfilter}
\alias{pfilter,pomp,numeric-method}
\title{Parallel particle filter computations}
\usage{
\S4method{pfilter}{pomp,numeric}(data, Nrep, ...)
}
\arguments{
\item{data}{either a data frame holding the time series data,
or an object of class \sQuote{pomp},
i.e., the output of another \pkg{pomp} calculation.}

\item{Nrep}{number of replicate particle filter computations to run.}

\item{...}{additional arguments supply new or modify existing model characteristics or components.
See \code{\link{pomp}} for a full list of recognized arguments.

When named arguments not recognized by \code{\link{pomp}} are provided, these are made available to all basic components via the so-called \dfn{userdata} facility.
This allows the user to pass information to the basic components outside of the usual routes of covariates (\code{covar}) and model parameters (\code{params}).
See the \link[=userdata]{userdata documentation here} for information on how to use this facility.}
}
\description{
Runs multiple instances of \code{pfilter} using \code{foreach}.
}
