% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{pfilter}
\alias{pfilter}
\alias{pfilter,ANY,numeric-method}
\alias{pfilter,ANY,missing-method}
\alias{pfilter,pompList,numeric-method}
\alias{pfilter,pompList,missing-method}
\title{Parallel particle filter computations}
\usage{
\S4method{pfilter}{ANY,numeric}(data, Nrep, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)

\S4method{pfilter}{ANY,missing}(data, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)

\S4method{pfilter}{pompList,numeric}(data, Nrep, ..., seed = TRUE, chunk.size = NULL, scheduling = 1)

\S4method{pfilter}{pompList,missing}(data, Nrep, ...)
}
\arguments{
\item{data}{passed to \code{\link[pomp:pfilter]{pomp::pfilter}}}

\item{Nrep}{number of replicate particle filter computations to run.
By default, \code{Nrep = 1}.}

\item{...}{all additional arguments are passed to \code{\link[pomp:pfilter]{pomp::pfilter}}}

\item{seed}{seed for the parallel random-number generator.
Setting \code{seed=TRUE} passes the current seed to the parallel RNG.
See \code{\link[doFuture]{\%dofuture\%}} for more information.}

\item{chunk.size}{average number of elements processed per chunk.
See \code{\link[doFuture]{\%dofuture\%}} for more information.}

\item{scheduling}{average number of chunks that each worker processes.
See \code{\link[doFuture]{\%dofuture\%}} for more information.
This is ignored unless \code{chunk.size=NULL}.}
}
\description{
Runs multiple instances of \code{pfilter} using \code{foreach}.
}
\examples{
library(circumstance)
library(doFuture)

ou2() -> ou2

plan(sequential)
system.time(ou2 |> pfilter(Np=10000,Nrep=6) -> pfs)

plan(multicore)
system.time(ou2 |> pfilter(Np=10000,Nrep=6) -> pfs)
}
\seealso{
\code{\link[pomp:pfilter]{pomp::pfilter}}.
}
