% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfilter.R
\name{pfilter}
\alias{pfilter}
\alias{pfilter,numeric-method}
\alias{pfilter-pomp,numeric}
\title{Parallel particle filter computations}
\usage{
\S4method{pfilter}{numeric}(Nrep, ...)
}
\arguments{
\item{Nrep}{number of replicate particle filter computations to run.}

\item{...}{all additional arguments are passed to \code{pomp::\link[pomp]{pfilter}}}
}
\description{
Runs multiple instances of \code{pfilter} using \code{foreach}.
}
\examples{
options(digits=3)

set.seed(789744859)

library(circumstance)
library(doParallel)
library(doRNG)

registerDoParallel()
registerDoRNG(789744859)

ou2() -> ou2
pfilter(ou2,Np=1000,Nrep=10) -> pfs
logmeanexp(logLik(pfs),se=TRUE)

registerDoSEQ()
}
\seealso{
\code{pomp::\link[pomp]{pfilter}}.
}
