% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{pmcmc}
\alias{pmcmc}
\alias{pmcmc,ANY,data.frame-method}
\alias{pmcmc,ANY,missing-method}
\alias{pmcmc,pompList,missing-method}
\alias{pmcmc,pfilterList,missing-method}
\alias{pmcmc,pmcmcList,missing-method}
\title{Particle Markov chain Monte Carlo in parallel}
\usage{
\S4method{pmcmc}{ANY,data.frame}(data, starts, ...)

\S4method{pmcmc}{ANY,missing}(data, starts, ...)

\S4method{pmcmc}{pompList,missing}(data, starts, ...)

\S4method{pmcmc}{pfilterList,missing}(data, starts, ...)

\S4method{pmcmc}{pmcmcList,missing}(data, starts, ...)
}
\arguments{
\item{data}{passed to \code{\link[pomp:pmcmc]{pomp::pmcmc}}}

\item{starts}{data frame containing parameters at which to begin iterated filtering}

\item{...}{all additional arguments are passed to \code{\link[pomp:pmcmc]{pomp::pmcmc}}}
}
\description{
Runs multiple instances of \code{pmcmc} using \code{foreach}.
}
\seealso{
\code{\link[pomp:pmcmc]{pomp::pmcmc}}.
}
